feat: shorten import paths
