+++ crates/bevy_mod_scripting_core/src/bindings/mod.rs
@@ -1,15 +1,15 @@
 //! Abstractions to help with creating bindings between bevy and scripting languages.
 
-pub mod access_map;
-pub mod allocator;
-pub mod function;
-pub mod globals;
-pub mod pretty_print;
-pub mod query;
-pub mod reference;
-pub mod schedule;
-pub mod script_system;
-pub mod script_value;
-pub mod world;
-
-pub use {allocator::*, query::*, reference::*, world::*};
+crate::private::export_all_in_modules!{
+    access_map,
+    allocator,
+    function,
+    globals,
+    pretty_print,
+    query,
+    reference,
+    schedule,
+    script_system,
+    script_value,
+    world,
+}
